# server/model/system/sys_base_menu_btn.go
package system

import "github.com/lindocedskes/model"

type SysBaseMenuBtn struct {
	model.BaseModel
	Name          string `json:"name" gorm:"comment:按钮关键key"`
	Desc          string `json:"desc" gorm:"按钮备注"`
	SysBaseMenuID uint   `json:"sysBaseMenuID" gorm:"comment:菜单ID"`
}